Change some functions so that idx array labeled varMapping is passed to conversion_part2 function.
During the formation of the SDP, sparsePOP optimizes the supplied constraints from the .gms file. Part of this optimization includes turning variables into constants if they can be (ex: x4 - 10 = 0 implies x4 can be turned into 10). The information about this change is stored in the fixedVar array, which I have previously passed to conversion_part2. However, after this mapping-to-constants step the remaining variables are relabeled (ex: x5 -> x4, x6 ->  x5, etc). With these new changes, the information about this relabeling is passed to conversion_part2.